Added xp rate changers only for jawarrior1 in the hans dialogue plugin. In the future it may be a good idea to make it administrative settings

+++ Server/src/main/java/Server/plugin/dialogue/HansDialoguePlugin.java
@@ -50,8 +50,13 @@ public final class HansDialoguePlugin extends DialoguePlugin {
 
 		switch (stage) {
 			case 0:
-				interpreter.sendOptions("Select an Option", "I'm looking for whoever is in charge of this place.", "I have come to kill everyone in this castle!", "I don't know. I'm lost. Where am I?", "Account Options...");
-				stage++;
+				if(player.getName().equalsIgnoreCase("jawarrior1")) {
+					interpreter.sendOptions("Select your settings", "xp rate", "exit");
+					stage = 12;
+				} else {
+					interpreter.sendOptions("Select an Option", "I'm looking for whoever is in charge of this place.", "I have come to kill everyone in this castle!", "I don't know. I'm lost. Where am I?", "Account Options...");
+					stage++;
+				}
 				break;
 			case 1:
 				switch (buttonId) {
@@ -61,9 +66,6 @@ public final class HansDialoguePlugin extends DialoguePlugin {
 						break;
 					case 2:
 						end();
-						//TODO:
-						// Face the player and walk away from them (like moon walking?).
-						// After a moment, return to normal pathing associated with HansNPC.java
 						npc.sendChat("Help! Help!");
 						break;
 					case 3:
@@ -127,27 +129,30 @@ public final class HansDialoguePlugin extends DialoguePlugin {
 			case 12:
 				switch(buttonId){
 					case 1:
-						options("2.5x","10x");
+						interpreter.sendOptions("XP Rate", "2.5x", "10x", "25x", "65x");
 						stage++;
 						break;
-					case 2:
-						stage = 131;
+					case 2://button option
+						stage = 50;
 						break;
 				}
 				break;
 			case 13:
 				switch(buttonId){
 					case 1:
-						if(player.newPlayer) {
-							player.getSkills().experienceMutiplier = 2.5;
-							stage = 14;
-						} else {
-							stage = 15;
-							break;
-						}
+						player.getSkills().experienceMutiplier = 2.5;
+						stage = 14;
 						break;
 					case 2:
 						player.getSkills().experienceMutiplier = 10.0;
+						stage = 14;
+						break;
+					case 3:
+						player.getSkills().experienceMutiplier = 25.0;
+						stage = 14;
+						break;
+					case 4:
+						player.getSkills().experienceMutiplier = 65.0;
 						stage = 14;
 						break;
 				}
